implemented deleting & downloading files

+++ app/Http/Controllers/FileController.php
@@ -2,14 +2,17 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\DestroyFilesRequest;
-use App\Http\Requests\StoreFileRequest;
-use App\Http\Requests\StoreFolderRequest;
-use App\Http\Resources\FileResource;
 use App\Models\File;
 use Inertia\Inertia;
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
+use App\Http\Resources\FileResource;
 use Illuminate\Support\Facades\Auth;
+use App\Http\Requests\StoreFileRequest;
+use Illuminate\Support\Facades\Storage;
+use App\Http\Requests\FilesActionRequest;
+use App\Http\Requests\StoreFolderRequest;
+use ZipArchive;
 
 class FileController extends Controller
 {
@@ -83,7 +86,7 @@ class FileController extends Controller
         // dd($data, $fileTree);
     }
 
-    public function destroy(DestroyFilesRequest $request)
+    public function destroy(FilesActionRequest $request)
     {
         $data = $request->validated();
         $parent = $request->parent;
@@ -97,11 +100,65 @@ class FileController extends Controller
         } else {
             foreach ($data['ids'] ?? [] as $id) {
                 $file = File::find($id);
-                $file->delete();
+                if ($file) {
+                    $file->delete();
+                }
             }
         }
 
         return to_route('myFiles', ['folder' => $parent->path]);
+    }
+
+    public function download(FilesActionRequest $request)
+    {
+        $data = $request->validated();
+        $parent = $request->parent;
+
+        $all = $data['all'] ?? false;
+
+        $ids = $data['ids'] ?? [];
+
+        if (!$all && empty($ids)) {
+            return [
+                'message' => 'Please select files to download'
+            ];
+        }
+
+        if ($all) {
+            $url = $this->createZip($parent->children);
+            $filename = $parent->name . '.zip';
+        } else {
+            if (count($ids) == 1) {
+                $file = File::find($ids[0]);
+                if ($file->is_folder) {
+                    if ($file->children->count() == 0) {
+                        return [
+                            'message' => 'This folder is empty.' 
+                        ];
+                    }
+
+                    $url = $this->createZip($file->children);
+                    $filename = $file->name . '.zip';
+
+                } else {
+                    $dest = 'public/' . pathinfo($file->storage_path, PATHINFO_BASENAME);
+                    Storage::copy($file->storage_path, $dest);
+
+                    $url = asset(Storage::url($dest));
+                    $filename = $file->name;
+                }
+            } else {
+                $files = File::query()->whereIn('id', $ids)->get();
+                $url = $this->createZip($files);
+
+                $filename = $parent->name . '.zip';
+            }
+        }
+
+        return [
+            'url' => $url,
+            'filename' => $filename,
+        ];
     }
 
     private function getRoot()
@@ -120,6 +177,38 @@ class FileController extends Controller
                 $this->saveFileTree($file, $folder, $user);
             } else {
                 $this->saveFile($file, $parent, $user);
+            }
+        }
+    }
+
+    public function createZip($files): string
+    {
+        $zipPath = 'zip/' . Str::random() . '.zip';
+        $publicPath = "public/$zipPath";
+
+        if (!is_dir(dirname($publicPath))) {
+            Storage::makeDirectory(dirname($publicPath));
+        }
+
+        $zipFile = Storage::path($publicPath);
+        $zip = new ZipArchive();
+
+        if ($zip->open($zipFile, ZipArchive::CREATE | ZipArchive::OVERWRITE) == true) {
+            $this->addFilesToZip($zip, $files);
+        }
+
+        $zip->close();
+
+        return asset(Storage::url($zipPath));
+    }
+
+    private function addFilesToZip($zip, $files, $ancestors = '')
+    {
+        foreach ($files as $file) {
+            if ($file->is_folder) {
+                $this->addFilesToZip($zip, $file->children, $ancestors . $file->name . '/');
+            } else {
+                $zip->addFile(Storage::path($file->storage_path), $ancestors . $file->name);
             }
         }
     }
